+++ lib/rodbot/plugins/word_of_the_day/schedule.rb
@@ -2,12 +2,14 @@
 
 require 'httpx'
 
+require_relative 'adapter'
+
 module Rodbot
   class Plugins
     class WordOfTheDay
       class Schedule
         def initialize
-          Clockwork.every(1.day, -> { Rodbot.say message }, at: time)
+          Clockwork.every(1.day, -> { Rodbot.say "Word of the day: #{message}" }, at: time)
         end
 
         private
@@ -16,36 +18,15 @@ module Rodbot
           Rodbot.config(:plugin, :word_of_the_day, :time) || '12:00'
         end
 
-        def message
-          Rodbot::Plugins::WordOfTheDay::Today.new.message
-        end
-
-      end
-
-      class Today
-        def initialize
-          @response = HTTPX.with(timeout: { request_timeout: 60 }).get('https://www.merriam-webster.com/word-of-the-day')
+        def languages
+          Rodbot.config(:plugin, :word_of_the_day, :languages) || %w(english)
         end
 
         def message
-          if @response.status == 200
-            "Word of the day: [#{word}](#{url})"
-          else
-            "Sorry, there was a problem fetching the word of the day."
-          end
-        end
-
-        private
-
-        def word
-          @response.body.to_s.match(/<h2 class="word-header-txt">(.+?)</).captures.first
-        end
-
-        def url
-          @response.body.to_s.match(/<meta property="og:url" content="(.+?)"/).captures.first
+          languages.map { Adapter.new(_1).message }.compact.join(' / ')
         end
       end
-
     end
   end
 end
+